created update-username-input, account-setting-form, notification-setting-form (#49)
* refactored account settings into update-username-input, account-setting-form, notification-setting-form
