+++ src/routes/signin/+page.svelte
@@ -3,5 +3,5 @@
 </script>
 
 <div class="flex h-screen w-full items-center justify-center px-4">
-	<SigninForm />
+  <SigninForm />
 </div>
