+++ src/routes/signup/form/sign-up-form.svelte
@@ -1,19 +1,21 @@
 <script lang="ts">
   import { Input } from '$lib/components/ui/input';
-  import * as Card from "$lib/components/ui/card"
-  import * as Form from "$lib/components/ui/form/index";
-  import * as InputOTP from "$lib/components/ui/input-otp/index.js";
-  import { REGEXP_ONLY_DIGITS } from "bits-ui";
+  import * as Card from '$lib/components/ui/card';
+  import * as Form from '$lib/components/ui/form/index';
+  import * as InputOTP from '$lib/components/ui/input-otp/index.js';
+  import { REGEXP_ONLY_DIGITS } from 'bits-ui';
   import { UserIdentType } from '@baragaun/bg-node-client';
   import { myUserContext } from '@/contexts/my-user-context.svelte';
-  import { formSchema, schemaStep1, schemaStep2, schemaStep3, type FormSchema } from './sign-up-form-schema';
-  import SuperDebug, {
-    type SuperValidated,
-    type Infer,
-    superForm,
-  } from "sveltekit-superforms";
-  import { zod, zodClient } from "sveltekit-superforms/adapters";
- 
+  import {
+    formSchema,
+    schemaStep1,
+    schemaStep2,
+    schemaStep3,
+    type FormSchema,
+  } from './sign-up-form-schema';
+  import SuperDebug, { type SuperValidated, type Infer, superForm } from 'sveltekit-superforms';
+  import { zod, zodClient } from 'sveltekit-superforms/adapters';
+
   let { data }: { data: { form: SuperValidated<Infer<FormSchema>> } } = $props();
 
   const form = superForm(data.form, {
@@ -46,24 +48,24 @@
   let currentStep = $state(1);
 
   // const { form, errors, message, enhance, validateForm, options } = superForm(data.form, {
-	// 	// No need for hidden fields with dataType: 'json'
-	// 	dataType: 'json',
+  // 	// No need for hidden fields with dataType: 'json'
+  // 	dataType: 'json',
   //   validators: zodClient(formSchema),
-	// 	async onSubmit({ cancel }) {
-	// 		// If on last step, make a normal request
-	// 		if (currentStep == steps.length) return;
-	// 		else cancel();
+  // 	async onSubmit({ cancel }) {
+  // 		// If on last step, make a normal request
+  // 		if (currentStep == steps.length) return;
+  // 		else cancel();
 
-	// 		// Make a manual client-side validation, since we have cancelled
-	// 		const result = await validateForm({ update: true });
-	// 		if (result.valid) currentStep = currentStep + 1;
-	// 	},
+  // 		// Make a manual client-side validation, since we have cancelled
+  // 		const result = await validateForm({ update: true });
+  // 		if (result.valid) currentStep = currentStep + 1;
+  // 	},
 
-	// 	async onUpdated({ form }) {
-	// 		if (form.valid) currentStep = 1;
-	// 	}
-	// });
- 
+  // 	async onUpdated({ form }) {
+  // 		if (form.valid) currentStep = 1;
+  // 	}
+  // });
+
   const { form: formData, enhance } = form;
   let email = $state('');
   let username = $state('');
@@ -75,7 +77,8 @@
   const validateEmail = async (email: string) => {
     const validFormat = /^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(email);
     if (validFormat) {
-      identIsAvailable = await myUserContext.isUserIdentAvailable(email, UserIdentType.email) || false
+      identIsAvailable =
+        (await myUserContext.isUserIdentAvailable(email, UserIdentType.email)) || false;
     }
     return identIsAvailable;
   };
@@ -92,7 +95,7 @@
     loading = true;
 
     try {
-      console.log('calling signup without data')
+      console.log('calling signup without data');
 
       // Actually sign up the User, then verify their email
       // const { myUser, error } = await myUserContext.signUp(email);
@@ -123,8 +126,8 @@
     loading = true;
 
     try {
-      console.log('trying onSignUp')
-      
+      console.log('trying onSignUp');
+
       // TODO: currently non functional until the flow is in place
 
       // Actually sign up the User, then verify their email
@@ -192,84 +195,85 @@
       </Card.Header>
       <Card.Content>
         <div class="grid gap-4">
-            <Form.Field {form} name="email">
-              <Form.Control>
-                {#snippet children({ props })}
-                  <Input
-                    {...props}
-                    id="email"
-                    type="email"
-                    placeholder="Email"
-                    required
-                    bind:value={$formData.email}
-                  />
-                {/snippet}
-              </Form.Control>
-              <Form.FieldErrors />
-            </Form.Field>
-            <Card.Description>
-              By continuing, you agree to our User Agreement and acknowledge that you understand the Privacy Policy.
-            </Card.Description>
-            <Form.Button
-              type="button"
-              class="w-full"
-              onclick={onSubmitEmail}
-              disabled={!$formData.email}
-            >
-              Continue
-            </Form.Button>
-            <div class="mt-4 text-center text-sm">
-              Already have an account?
-              <a href="/signin" class="underline"> Sign in </a>
-            </div>
+          <Form.Field {form} name="email">
+            <Form.Control>
+              {#snippet children({ props })}
+                <Input
+                  {...props}
+                  id="email"
+                  type="email"
+                  placeholder="Email"
+                  required
+                  bind:value={$formData.email}
+                />
+              {/snippet}
+            </Form.Control>
+            <Form.FieldErrors />
+          </Form.Field>
+          <Card.Description>
+            By continuing, you agree to our User Agreement and acknowledge that you understand the
+            Privacy Policy.
+          </Card.Description>
+          <Form.Button
+            type="button"
+            class="w-full"
+            onclick={onSubmitEmail}
+            disabled={!$formData.email}
+          >
+            Continue
+          </Form.Button>
+          <div class="mt-4 text-center text-sm">
+            Already have an account?
+            <a href="/signin" class="underline"> Sign in </a>
           </div>
-        </Card.Content>
-      {:else if currentStep === 2}
-        <Card.Header>
-          <Card.Title class="text-2xl">Sign Up</Card.Title>
-        </Card.Header>
-        <Card.Content>
-          <div class="grid gap-4">
-            <Form.Field {form} name="emailOtp">
-              <Form.Control>
-                {#snippet children({ props })}
-                  <InputOTP.Root
-                    {...props}
-                    id="emailOtp"
-                    maxlength={6}
-                    pattern={REGEXP_ONLY_DIGITS}
-                    required
-                    bind:value={$formData.emailOtp}
-                  >
-                    {#snippet children({ cells })}
-                      <InputOTP.Group>
-                        {#each cells as cell (cell)}
-                          <InputOTP.Slot {cell} />
-                        {/each}
-                      </InputOTP.Group>
-                    {/snippet}
-                  </InputOTP.Root>
-                {/snippet}
-              </Form.Control>
-              <Form.Description>Enter the six digit code sent to your email.</Form.Description>
-              <Form.FieldErrors />
-            </Form.Field>
-            <Form.Button
-              type="button"
-              class="w-full"
-              onclick={onSubmitEmailOtp}
-              disabled={!$formData.emailOtp}
-            >
-              Continue
-            </Form.Button>
-            <div class="mt-4 text-center text-sm">
-              Already have an account?
-              <a href="/signin" class="underline"> Sign in </a>
-            </div>
-            </div>
-          </Card.Content>
-      {/if}
+        </div>
+      </Card.Content>
+    {:else if currentStep === 2}
+      <Card.Header>
+        <Card.Title class="text-2xl">Sign Up</Card.Title>
+      </Card.Header>
+      <Card.Content>
+        <div class="grid gap-4">
+          <Form.Field {form} name="emailOtp">
+            <Form.Control>
+              {#snippet children({ props })}
+                <InputOTP.Root
+                  {...props}
+                  id="emailOtp"
+                  maxlength={6}
+                  pattern={REGEXP_ONLY_DIGITS}
+                  required
+                  bind:value={$formData.emailOtp}
+                >
+                  {#snippet children({ cells })}
+                    <InputOTP.Group>
+                      {#each cells as cell (cell)}
+                        <InputOTP.Slot {cell} />
+                      {/each}
+                    </InputOTP.Group>
+                  {/snippet}
+                </InputOTP.Root>
+              {/snippet}
+            </Form.Control>
+            <Form.Description>Enter the six digit code sent to your email.</Form.Description>
+            <Form.FieldErrors />
+          </Form.Field>
+          <Form.Button
+            type="button"
+            class="w-full"
+            onclick={onSubmitEmailOtp}
+            disabled={!$formData.emailOtp}
+          >
+            Continue
+          </Form.Button>
+          <div class="mt-4 text-center text-sm">
+            Already have an account?
+            <a href="/signin" class="underline"> Sign in </a>
+          </div>
+        </div>
+      </Card.Content>
+    {/if}
   </Card.Root>
-  
+
   <!-- <SuperDebug data={form} /> -->
 </form>
